Very improved the splash page.
