+++ coach-me/src/components/coach/coachRegistration/coachRegistration.scss
@@ -9,7 +9,7 @@
 
 .Register-Wrapper {
     display: flex;
-    height: 100vh;
+    height: 100%;
     font-family: DM Sans;
 
     .side-one {
@@ -23,7 +23,7 @@
         justify-content: space-between;
 
         .objective {
-            font-family: DM Sans;
+            font-family: 'DM Sans', serif;
             font-style: italic;
             font-weight: 500;
             font-size: 2rem;
