+++ coach-me/src/components/splash/AboutUs.jsx
@@ -1,38 +1,27 @@
 import React,{useState} from 'react';
 import { Link } from 'react-router-dom';
+import Login from './Login';
+
 import './splash.scss';
+
 
 const AboutUs = () => {
 
-    const [showCoachLogin, setshowCoachLogin] =  useState(false)
-    const clickHandler = () => {
-        setshowCoachLogin(!showCoachLogin);
-    }
+
     return (
         <div className='about'>
-            <h4>About CoachMe</h4>
-            <p>CoachMe offers tele-health coaching to individuals with cardiovascular disease (CVD).  We match real, trained coaches with patients to help them receive the personalized support they need to make a change.</p>
-
-            <p className="speak-to-rep">If you'd like more information <a href="">Speak to a representative</a> to learn more! If you're ready to get started, use the buttons below! </p>
-
-            <div className="buttons">
-                <div className={showCoachLogin ? "clients-buttons hide" : "clients-buttons button-box"}>
-                    <Link to='/client-login'><button>Log in</button></Link>
-                    <Link to='/createAccount'><button>Sign up</button></Link>
-                    
-
-                    <p className='coach-toggle'
-                    onClick={clickHandler}
-                    >Show me Coach options.</p>
-                </div>
-                <div className={showCoachLogin ? "coaches-buttons button-box" : "coaches-buttons hide"}>
-                <Link to='/coach-login'><button>Log in as Coach</button></Link>
-                <Link to='/coach-register'><button>Apply to be a Coach</button></Link>
-                    <p className='coach-toggle'
-                    onClick={clickHandler}
-                    >Show me Client options.</p>
-                </div>
+          <div className="how-we-do">
+              <h4 className="about-header">How we do it</h4>
+              <p className="about-text bold">We believe in the power of people to drive lifestyle changes and better outcomes.</p>
+              <p className="about-text">Our tech-enabled platform allows health coaches to reach patients where they are every day - on their phones.</p>
+            </div>  
+            <div className="impact">
+              <h4 className="about-header">Our Impact</h4>
+              <p className="about-text bold">Medicaid pays billions of dollars a year on treating downstream effects of diabetes and heart disease.</p>
+              <p className="about-text">But these conditions are reversible. Health coaching has shown sustained improvements in health outcomes specifically in low-income populations</p>
             </div>
+
+            <Login />
         </div>
     );
 }
