+++ coach-me/src/components/splash/Splash.jsx
@@ -8,14 +8,17 @@ const Splash = () => {
         <div className='splash-screen'>
             <header>
                 <div className="hero-text">
-                <h1>This is Coach me, dawg.</h1>
-                <p>We here to halp.</p>
+                <h1>Coach Me.</h1>
+                <p>We're a <span>patient-first</span> non-profit fighting <span>chronic disease</span>.</p>
+                <p className='small-text'>Improving the health of low-income Americans with diabetes and heart disease with health coaching for those who need it most.</p>
                 </div>
+                
                 <div className="hero-img">
                     <img src={happy} alt="A happy Woman in Fall"/>
                 </div>
             </header>
             <AboutUs />
+            
             
         </div>
     );
