+++ coach-me/src/components/splash/splash.scss
@@ -1,6 +1,11 @@
+@import url('https://fonts.googleapis.com/css?family=Poppins&display=swap');
+@import url('https://fonts.googleapis.com/css?family=DM+Sans&display=swap');
+
 $dark-background: #274160;
-$light-background: rgba(100, 209, 252, 0.5);
+$light-background: #355478;
 $font-color: white;
+$header-font: 'Poppins', sans-serif;
+$text-font: 'DM Sans', serif;
 
 .splash-screen {
     display: flex;
@@ -11,28 +16,39 @@ $font-color: white;
     background-color: $dark-background;
     height: 100%;
     header {
+        border-radius: 8px;
         background-color: $light-background;
         padding: 2%;
         display: flex;
         justify-content: space-evenly;
         align-items: center;
-        width: 80%;
+        width: 90%;
         margin: 0 auto;
         .hero-text {
             width: 50%;
+            text-shadow: 0px 0px 2px #333;
 
             h1 {
-                font-size: 5em;
+                font-size: 7em;
                 text-align: center;
                 color: $font-color;
+                font-family: $header-font;
             }
             p {
-                font-size: 3em;
+                font-size: 3.3em;
                 text-align: center;
                 margin-top: 1em;
                 color: $font-color;
+                font-family: $text-font;
+                span {
+                    color: #ff98a2;
+                }
+            }
+            .small-text {
+                font-size: 1.2em;
             }
         }
+
         .hero-img {
             margin-left: 3em;
 
@@ -43,35 +59,66 @@ $font-color: white;
 
             width: 50%;
         }
+        @media (max-width: 800px) {
+            flex-direction: column;
+            justify-content: center;
+            align-items: center;
+            .hero-text {
+                width: 80%;
+                text-shadow: 0px 0px 2px #333;
+            }
+        }
     }
 
     .about {
         display: flex;
+        border-radius: 8px;
         flex-direction: column;
         justify-content: center;
         align-items: center;
         background-color: white;
-        width: 95%;
+        width: 90%;
         padding: 2%;
-        margin: 1em auto;
-        h4 {
-            font-size: 2.6em;
+        margin: 2em auto;
+        .about-header {
+            font-size: 2.8em;
             font-weight: bold;
-            margin-bottom: 1em;
+            margin-bottom: 0.7em;
+            font-family: $header-font;
+            margin-left: 0.5em;
+            color: $dark-background;
         }
-        p {
-            font-size: 1.8em;
+        .about-text {
+            font-size: 1.6em;
+            margin: 1em 0;
+            font-family: $text-font;
+            text-align: center;
+            line-height: 1.2em;
         }
-        .speak-to-rep {
-            margin-top: 2em;
+        .bold {
+            font-weight: bold;
         }
+
         .buttons {
             width: 90%;
             margin: 1em auto;
+            .speak-to-rep {
+                margin-top: 2em;
+                font-size: 1.1em;
+
+                a {
+                    color: dodgerblue;
+                    text-shadow: 0px 0px 1px #333;
+                }
+            }
             .button-box {
                 transition: all 0.3s ease;
                 button {
+                    font-family: $header-font;
                     width: 100%;
+                    margin-top: 1.1em;
+                    font-size: 1.3em;
+                    box-shadow: 1px 1px 2px #333;
                 }
 
                 p {
@@ -80,6 +127,7 @@ $font-color: white;
                     margin-top: 1em;
                     text-decoration: underline;
                     cursor: pointer;
+                    font-family: $text-font;
                 }
             }
         }
